+++ js/javascript.js
@@ -1,6 +1,7 @@
 
 const DEBUG = false;
 //=include util/util.js
+//=include util/_electron.js
 //=include _load.js
 //=include _data.js
 //=include _sound.js